# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apartment Watchlist API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingChange } from "./listingChange";

export interface DashboardSummary {
  totalListings: number;
  activeListings: number;
  priceDropsToday: number;
  statusChangesToday: number;
  uncheckedListings: number;
  unavailableListings: number;
  unreadNotifications: number;
  recentChanges: ListingChange[];
}