# lib/api-zod/src/generated/types/listingSnapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Apartment Watchlist API
 * OpenAPI spec version: 0.1.0
 */

export interface ListingSnapshot {
  id: number;
  listingId: number;
  checkedAt: string;
  /** @nullable */
  extractedData?: string | null;
  /** @nullable */
  listingStatus?: string | null;
  /** @nullable */
  currentPrice?: string | null;
  fetchSuccess: boolean;
  /** @nullable */
  errorMessage?: string | null;
}